013.24793845-ignoreascii.patch
This patch is a backport of proftpd Bug#4159 from 1.3.6.rc2 to 1.3.5b. It should
be removed as soon as we sync with stable 1.3.6 release.

This patch is hand crafted combination of following two commits. No test files
are included.

Fix regression, caused by change for Bug#4159 with the ASCII translation handling of LIST/NLST output
https://github.com/proftpd/proftpd/commit/3892250b429661dd8d166a55230f9a0da923a26b

Bug#4159: Support ability to disable ASCII translation transparently to FTP clients.
https://github.com/proftpd/proftpd/commit/68229c0bc839260a43ac1c9b9f7ed68e39aa0f1a

http://bugs.proftpd.org/show_bug.cgi?id=4159